Update codebase to use VFS instead of the hash
I had been expecting to use an interface very similar to a hash all
along, and it worked out pretty well. The main update I needed to do was
to fully remove `outfiles` as a concept and merge it in with the single
unifying VFS.

The original reason I had outfiles as separate was to make testing a
little more straightforward (since outfiles would only contain modified
files, I wouldn't need to check for as many cases). However the change
to using the VFS didn't make things much harder at all.