Configure to run right out of the box if Rails project
The CLI help screen before suggested to create a config file in Rails
applications and specified what the config file should be.

I realized that it'd be much easier if things worked out of the box for
Rails. Since I knew exactly what configs were required for Rails apps
(the necessary configuration was already printed in the help screen),
it was fairly easy to auto-merge the Rails configs into the defaults.

I also noticed that `Gem::Text` wasn't always loaded depending on how
the executable was called, so I've required it explicitly.