Fix Rails 6 support
This doesn't yet support Rails 6's multiple databases.

Rails 6 renamed `#parent` and `#parents` with `#module_parent` and
`#module_parents`, respectively, and added an alias with a deprecation
notice.

I updated Ahnnotate to use the Rails 6 method names (`#module_parent`),
and similar to it I added an alias for compatibility with Rails 4 and 5.